requirements.txt
streamlit==1.35.0
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.1
